fix: wait for cursor controls on accepted macOS sockets
